First Assignment
Huzzah! I have completed week 1 of 4 in the class! I was tasked with creating a tip calculator for 15%, 18%, and 20% from a final bill in Python. For extra credit, I replaced the "$" sign so that it would not break my code, and even added it to the final product to make readability easier. Ain't much, but it's honest work!