Functions Challenge
In this challenge we were tasked with creating an Uppercase and Reverse function, in which we would take a small amount of text (EG: banana), and change it so every letter is capitalized and the word (or words) are reversed. I think there is probably an easier way to do this, but this is where googling got me and I came out with an answer that worked so that's a win in my book!