Week 2 Assignment
Errors.py is my Week 2 assignment from One Month, where I was supposed to debug the code that was given to me (original_errors.py). This was the hardest challenge yet, especially since the original line 61 if errors.py was print(sing(99)), which was causing an additional None to print out at the end. This took me much longer than I care to admit to solve. But I solved it right? Right. Code debugged, on to greater things!